refactor: move shared color, index to constants
